requirements.txt
streamlit
PyPDF2
python-docx
beautifulsoup4
markdown
langchain
langchain-community
langchain_google_genai
google-generativeai
faiss-cpu
python-dotenv


